requirements.txt
google-genai
mypy
types-requests
dotenv